requirements.txt
ajaxuploader==0.3.8
arcgis-rest-query==0.14
certifi==2017.4.17
chardet==3.0.3
diff-match-patch==20121119
Django==1.11
django-allauth==0.31.0
django-autoslug==1.9.3
django-crispy-forms==1.6.1
django-debug-toolbar==1.4
django-endless-pagination==2.0
django-filter==1.0.2
django-formtools==1.0
django-import-export==0.4.4
django-ipware==1.0.0
django-localflavor==1.1
django-passwords==0.2.0
django-queryset-csv==0.3.3
django-read-only-admin==0.1.2
django-tables2==1.5.0
django-tables2-reports==0.0.10
easy-thumbnails==2.5
et-xmlfile==1.0.1
idna==2.5
jdcal==1.3
num2words==0.5.3
numpy==1.9.2
oauthlib==2.0.2
odfpy==1.3.6
olefile==0.44
openpyxl==2.4.9
Pillow==4.0.0
pkg-resources==0.0.0
psycopg2==2.6
pyClamd==0.3.16
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.2
PyYAML==3.12
requests==2.17.3
requests-oauthlib==0.8.0
simplejson==3.6.5
six==1.9.0
sqlparse==0.1.18
stripe==1.58.0
tablib==0.11.2
unicodecsv==0.14.1
urllib3==1.21.1
vectorformats==0.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==0.7.5
